+++ feature/local/src/main/kotlin/uk/govuk/app/local/data/LocalRepo.kt
@@ -10,8 +10,8 @@ import uk.govuk.app.local.data.remote.model.RemoteLocalAuthority
 import uk.govuk.app.local.data.remote.safeLocalApiCall
 import uk.govuk.app.local.domain.model.Address
 import uk.govuk.app.local.domain.model.LocalAuthority
-import uk.govuk.app.local.toAddress
-import uk.govuk.app.local.toLocalAuthority
+import uk.govuk.app.local.domain.toAddress
+import uk.govuk.app.local.domain.toLocalAuthority
 import javax.inject.Inject
 import javax.inject.Singleton
