+++ feature/local/src/main/kotlin/uk/govuk/app/local/domain/LocalAuthorityMapper.kt
@@ -1,4 +1,4 @@
-package uk.govuk.app.local
+package uk.govuk.app.local.domain
 
 import uk.govuk.app.local.data.local.model.StoredLocalAuthority
 import uk.govuk.app.local.data.remote.model.RemoteAddress
